Move most of the logic from ApprovedPeer to stage::Approved
